fixed merge conflict and error in console
